EF: various warnings from iOS/clang

+++ iModelCore/DgnPlatform/Tests/DgnProject/NonPublished/TransactionManager_Test.cpp
@@ -249,9 +249,9 @@ void TxnMonitorVerifier::_OnTxnBoundary(TxnSummaryCR summary)
         auto eid = stmt.GetValueId<DgnElementId>(0);
         switch (stmt.GetValueInt(1))
             {
-            case TxnSummary::ChangeType::Add:    m_adds.insert(eid); break;
-            case TxnSummary::ChangeType::Delete: m_deletes.insert(eid); break;
-            case TxnSummary::ChangeType::Update: m_mods.insert(eid); break;
+            case (int)TxnSummary::ChangeType::Add:    m_adds.insert(eid); break;
+            case (int)TxnSummary::ChangeType::Delete: m_deletes.insert(eid); break;
+            case (int)TxnSummary::ChangeType::Update: m_mods.insert(eid); break;
             default:
                 FAIL();
             }
